backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.23.0
tensorflow>=2.13.0
ultralytics>=8.0.0
